Refactor CChip class to a separate new CChip.cs file for DTXMania.csproj Replace all CDTX.CChip to CChip in other source files in DTXMania.csproj
